Added comments and removed old code
