+++ components/StructureAnimation.js
@@ -1,193 +1,128 @@
 import React, {Component} from 'react';
 import treeData from './../AnimationData/treeStructure';
+import Tree from './Tree';
+import Link from './Link';
 
-var d3 = require('d3');
-var ReactDOM = require('react-dom');
-var _ = require('lodash');
+const d3 = require('d3');
+const ReactDOM = require('react-dom');
 
-// StructureAnimation is based on this blog post:
-// http://javascript.tutorialhorizon.com/2014/09/08/render-a-d3js-tree-as-a-react-component/.
-// The problem here is that D3 is creating DOM elements without React's knowledge, using the renderTree component at the bottom of this page.
+// Note from Isaac: My goal her was to make React responsible for DOM structure (adding and removing elements) and to make D3 responsible for styling, as described in this blog post:
+// https://medium.com/@sxywu/on-d3-react-and-a-little-bit-of-flux-88a226f328f3#.ztcxqykek
 
 export default class StructureAnimation extends Component {
 
   constructor(props) {
     super(props);
     this.state = {
-      treeData: props.initialTreeData
+      treeData: props.initialTreeData,
+      margin: props.initialMargin,
+      windowWidth: props.initialWindowWidth,
+      windowHeight: props.initialWindowHeight
     }
   }
 
-  componentDidMount(){
-    var mountNode = ReactDOM.findDOMNode(this.refs.treeRender);
-    // Render the tree usng d3 after first component mount
-    renderTree(this.state.treeData, mountNode);
+  componentDidMount() {
+    // This fires on initial load and when I toggle between the structure and Git animations.
+    // For some reason, after I toggle, every time I run a command or toggle again, I see an error message in the console about setting state for an unmounted component:
+    // "Warning: setState(...): Can only update a mounted or mounting component. This usually means you called setState() on an unmounted component. This is a no-op. Please check the code for the undefined component."
+    // I'm not sure what's going on here, since componentDidMount shouldn't fire for an unmounted component.
+    // However, the tree still renders correctly.
+    ipcRenderer.send('ready-for-schema', '\n');
+    ipcRenderer.on('direc-schema', (e,arg)=>{
+      this.updateTree(arg);
+    });
   }
 
-  shouldComponentUpdate(nextProps, nextState){
-    if (_.isEqual(this.lastState,nextState)){
-      return false
-    } else {
-      // Delegate rendering the tree to a d3 function on prop change
-      renderTree(nextState.treeData, ReactDOM.findDOMNode(this.refs.treeRender));
-      // Do not allow react to render the component on prop change
-      return false;
-    }
-  }
-
-  updateTree(newSchema){
-    this.lastState = this.state
+  updateTree(newSchema) {
     this.setState({
-      treeData: newSchema
-    })
+      treeData: newSchema,
+      windowHeight: window.innerHeight,
+      windowWidth: window.innerWidth
+    });
   }
 
   render() {
-    ipcRenderer.on('direc-schema', (e,arg)=>{
-      this.updateTree(arg);
-    })
+    // Create variables to determine the size of the tree and the size of the SVG containing it (or just the height-width ratio?).
+    var viewBoxWidth = this.state.windowWidth * 7 / 12; // previously hard-coded as 660
+    var viewBoxHeight = this.props.sidebarVisible ? this.state.windowHeight * 7 / 24 : this.state.windowHeight * 5 / 24; // previously hard-coded as 300, which looks good if the sidebar and Chrome dev tools are visible
 
-    return (
+    // Create a tree layout.
+    var tree = d3.layout.tree()
+    // The first argument below is the maximum x-coordinate D3 will assign.
+    // The second argument is the maximum y-coordinate.
+    // We're switching width and height here because d3 by default makes trees that branch vertically, and we want a tree that branches horizontally.
+    // In other words, nodes that are on the same level will have the same y-coordinate but different x-coordinates.
+      .size([viewBoxHeight * 0.9, viewBoxWidth  * 0.9]);
+
+    // We know that the first node in the array is the root of the tree. Let's designate its initial coordinates - where it should enter.
+    var root = this.state.treeData[0];
+    root.x0 = viewBoxHeight / 2;
+    root.y0 = 0;
+
+    // The next line creates and returns an array of nodes associated with the specified root node. (The returned array is basically a flattened version of treeData.)
+    // Before the next line runs, each node has children, level, name, and value properties.
+    // After the next line runs, each node also has parent, depth, x, and y properties. (D3 tree nodes always have parent, child, depth, x, and y properties.)
+    // We will pass one node from this array to each Tree as props.data.
+    var nodes = tree.nodes(root),
+      // This line creates and returns an array of objects representing all parent-child links in the nodes array we just created.
+      linkSelection = tree.links(nodes);
+
+    nodes.forEach(function(d) {
+      // The default y-coordinates provided by d3.tree will make the tree stretch all the way across the screen.
+      // We want to compress the tree a bit so that there's room for file/directory names to the right of the deepest level.
+      d.y *= .8;
+      // We could scale d.x too to prevent the svg from cutting off notes at the top and bottom of a node column.
+      // If the node has a parent, set the node's initial coordinates to the parent's initial coordinates.
+      // In other words, the parent and child should enter from the same place.
+      if (d.parent) {
+        d.x0 = d.parent.x0;
+        d.y0 = d.parent.y0;
+      }
+    });
+
+    // I think it makes sense to the use target.name as an id, because no two links should ever point to the same target.
+    // If needed, though, we could use {link.source.name + '/'  link.target.name} instead.
+    // We sometimes receive directory names with slashes and sometimes receive them without, so I'm removing any slash before using the name as a key.
+    // I added the trim to account for the fact that the name values sometime begin with a carriage return for some reason, and that throws things off.
+    var links = linkSelection && linkSelection.map((link) => {
+      link.target.name = link.target.name.trim();
+      var nameEndsWithSlash = link.target.name.indexOf('/') === link.target.name.length - 1;
+      var key = nameEndsWithSlash ? link.target.name.slice(0, link.target.name.length - 1) : link.target.name;
+      return (<Link key={link.target.name} data={link} />)
+    });
+
+    var trees = nodes && nodes.map((node) => {
+      node.name = node.name.trim();
+      var nameEndsWithSlash = node.name.indexOf('/') === node.name.length - 1;
+      var key = nameEndsWithSlash ? node.name.slice(0, node.name.length - 1) : node.name;
+      return (<Tree key={key} data={node} />);
+    });
+
+    var viewBoxString = `0 0 ${viewBoxWidth} ${viewBoxHeight}`;
+
+    var translationValue = `translate(${this.state.margin.left}, ${this.state.margin.top})`;
+
+    // If you want to see the size of the SVG, add this code before the links and trees:
+    // <rect x='0' y='0' width={viewBoxWidth - this.state.margin.left} height={viewBoxHeight - this.state.margin.top} rx='15' ry='15' />
+    // How does the svg know to fill the full width of its containing div? I'm not sure about this.
+    return(
       <div id='Structure-Animation'>
-        <svg ref="treeRender"></svg>
+        <svg viewBox={viewBoxString}>
+          <g transform={translationValue}>
+            {links}
+            {trees}
+          </g>
+        </svg>
       </div>
     )
   }
 }
 
 StructureAnimation.defaultProps = {
-  initialTreeData: treeData
-}
-
-var renderTree = function(treeData, svgDomNode) {
-
-  var margin = {top: 0, right: 90, bottom: 0, left: 180},
-  	width = 660 - margin.right - margin.left,
-  	height = 200 - margin.top - margin.bottom;
-
-    var i = 0,
-      duration = 450,
-      root;
-
-    // Cleans up the SVG on re-render
-    d3.select(svgDomNode).selectAll("*").remove();
-
-    var tree = d3.layout.tree()
-      .size([height, width]);
-
-    var diagonal = d3.svg.diagonal()
-      .projection(function(d) { return [d.y, d.x]; });
-
-    var svg = d3.select(svgDomNode)
-      .attr("width", width + margin.right + margin.left)
-      .attr("height", height + margin.top + margin.bottom)
-      .append("g")
-      .attr("transform", "translate(" + margin.left + "," + margin.top + ")");
-
-    root = treeData[0];
-    root.x0 = height / 2;
-    root.y0 = 0;
-
-    update(root);
-
-    function update(source) {
-      // Compute the new tree layout.
-      var nodes = tree.nodes(root).reverse(),
-        links = tree.links(nodes);
-
-      // Normalize for fixed-depth.
-      nodes.forEach(function(d) { d.y = d.depth * 180; });
-
-      // Update the nodes…
-      var node = svg.selectAll("g.node")
-        .data(nodes, function(d) { return d.id || (d.id = ++i); });
-
-      // Enter any new nodes at the parent's previous position.
-      var nodeEnter = node.enter().append("g")
-        .attr("class", "node")
-        .attr("transform", function(d) { return "translate(" + source.y0 + "," + source.x0 + ")"; })
-        .on("click", click);
-
-      nodeEnter.append("circle")
-        .attr("r", 1e-6)
-        .style("fill", function(d) { return d._children ? "lightsteelblue" : d.level; });
-
-      nodeEnter.append("text")
-        .attr("x", function(d) { return d.children || d._children ? -20 : 20; })
-        .attr("dy", ".35em")
-        .attr("text-anchor", function(d) { return d.children || d._children ? "end" : "start"; })
-        .text(function(d) { return d.name; })
-        .style("fill-opacity", 1e-6);
-
-      // Transition nodes to their new position and make them visible.
-      // We need both nodeEnter and nodeUpdate in order to see the nodes (without nodeEnter, nothing is appended.)
-      var nodeUpdate = node.transition()
-        .duration(duration)
-        .attr("transform", function(d) { return "translate(" + d.y + "," + d.x + ")"; });
-
-      nodeUpdate.select("circle")
-        .attr("r", function(d) { return d.value ? d.value : 5; })
-        .style("fill", function(d) { return d._children ? "lightsteelblue" : d.level; });
-
-      nodeUpdate.select("text")
-        .style("fill-opacity", 1);
-
-      // Transition exiting nodes to the parent's new position.
-      // I believe this is used only when the user clicks on a parent node to hide its children.
-      var nodeExit = node.exit().transition()
-        .duration(duration)
-        .attr("transform", function(d) { return "translate(" + source.y + "," + source.x + ")"; })
-        .remove();
-
-      nodeExit.select("circle")
-        .attr("r", 1e-6);
-
-      nodeExit.select("text")
-        .style("fill-opacity", 1e-6);
-
-      // Update the links…
-      var link = svg.selectAll("path.link")
-        .data(links, function(d) { return d.target.id; });
-
-      // Enter any new links at the parent's previous position.
-      link.enter().insert("path", "g")
-        .attr("class", "link")
-        .attr("d", function(d) {
-        var o = {x: source.x0, y: source.y0};
-        return diagonal({source: o, target: o});
-        });
-
-      // Transition links to their new position.
-      link.transition()
-        .duration(duration)
-        .attr("d", diagonal);
-
-      // Transition exiting nodes to the parent's new position.
-      link.exit().transition()
-        .duration(duration)
-        .attr("d", function(d) {
-        var o = {x: source.x, y: source.y};
-        return diagonal({source: o, target: o});
-        })
-        .remove();
-
-      // Stash the old positions for transition.
-      nodes.forEach(function(d) {
-      d.x0 = d.x;
-      d.y0 = d.y;
-      });
-    }
-
-    // Toggle children on click.
-    function click(d) {
-      // d is an object - what you clicked
-      if (d.children) {
-      d._children = d.children;
-      d.children = null;
-      } else {
-      d.children = d._children;
-      d._children = null;
-      }
-      update(d);
-    }
+  initialTreeData: treeData,
+  // We're not currently using the right or bottom margins.
+  initialMargin: {top: 8, right: 20, bottom: 0, left: 90},
+  // The initial window dimensions are specified in app.on('ready') in main.js.
+  initialWindowWidth: 1200,
+  initialWindowHeight: 700
 }
